80% pagination

+++ eport-system/src/main/java/vn/com/irtech/eport/equipment/mapper/EquipmentDoMapper.java
@@ -86,6 +86,6 @@ public interface EquipmentDoMapper
 
   public List<EquipmentDo> getContainerListByIds(String[] ids);
 
-  public int getTotalPages();
+  public Long getTotalPages();
 	
 }
