+++ eport-system/src/main/java/vn/com/irtech/eport/equipment/service/impl/EquipmentDoServiceImpl.java
@@ -144,7 +144,7 @@ public class EquipmentDoServiceImpl implements IEquipmentDoService
     }
 
     @Override
-    public int getTotalPages()
+    public Long getTotalPages()
     {
         return equipmentDoMapper.getTotalPages();
     }
